Find new mirrors successfully
I had to visualize it in order to get the glyphs lined up for comparison.
To visualize it, I rewrote it to do an xor composite operation on a single canvas instead of comparing separate image data.
The trick was using textAlign instead of translating by the canvas width, so it effectively uses the glyph's center. I could also use textAlign = "center" presumably.